Add files via upload
Assignment Problems for Demonstration

1. Move the robot using keyboard control and incorporate distance check:
    Preliminary Directions: 
-- create a new folder in the Module2 folder you had created for Exercise 9 and name it 
Assignments
-- create a new file in this folder and name it problem1.py
-- copy the code from compass_robot.py to problem1.py
--  integrate code from distance_values.py to problem1.py

Expected output: The robot must be controlled by the user’s keyboard. The distance values (to the nearest obstacle sensed by the distance sensor) must be displayed on the console ONLY when the robot is MOVING but NOT WHEN STOPPED.

Notes: You can also access compass_robot.py from here: https://github.com/DexterInd/GoPiGo3/blob/master/Projects/CompassRobot/compass_robot.py
-------------------------------------------------------------------------------------------------------------------------------


2. Move robot using keyboard control, incorporate distance check and capture a photo when distance is below a set threshold
  Preliminary Directions: 
-- go to Assignments folder in Module 2 folder
-- create a new file in this folder and name it problem2.py
-- copy the code from problem1.py
-- integrate camera photo capture into problem1.py
-- set a conditional check: IF distance < 500mm then CAPTURE photo
			
Expected output: The robot must be controlled by the user’s keyboard. The distance values (to the nearest obstacle sensed by the distance sensor) must be displayed on the console ONLY when the robot is MOVING but NOT WHEN STOPPED
AND
The robot must capture a picture every time the robot senses an obstacle within 500mm. The photos must be named as module2_problem2_1, module2_problem2_2, module2_problem2_3, and so on..